// packages/meta/src/builder/node-types.ts
export enum NodeType {
  Object = "object",
  ObjectProperty = "object-property",
  ObjectPropertyDependency = "object-property-dependency",
  Predicate = "predicate",
  Operator = "operator",
  Array = "array",
  Grid = "grid",
  Enum = "enum",
  MultiEnum = "multi-enum",
  EnumItem = "enum-item",
  String = "string",
  Number = "number",
  Boolean = "boolean",
  File = "file",
  Tags = "tags",
  Range = "range",
}

export enum RangeValueType {
  String = "string",
  Number = "number",
}